WIP - refactor events and props into modules
