+++ package.json
@@ -16,6 +16,7 @@
   "author": "Kevin Yank <kevin@cultureamp.com>",
   "license": "MIT",
   "devDependencies": {
+    "@rollup/plugin-node-resolve": "^9.0.0",
     "@stencil/core": ">=1.8.0",
     "@types/node": "^14.0.13",
     "rollup": "^2.16.1",
