+++ rollup.config.js
@@ -1,10 +1,11 @@
 import pkg from './package.json';
 import sourcemaps from 'rollup-plugin-sourcemaps';
+import nodeResolve from '@rollup/plugin-node-resolve';
 
 export default {
   input: 'dist/index.js',
 
-  plugins: [sourcemaps()],
+  plugins: [sourcemaps(), nodeResolve()],
 
   external: ['path', 'node-sass', 'fs', 'util'],
