+++ src/props/unsupported-prop.ts
@@ -1,0 +1,7 @@
+import { Prop } from './prop';
+
+export class UnsupportedProp extends Prop {
+  isSupported(): boolean {
+    return false;
+  }
+}
